requirements.txt
google-generativeai 
python-dotenv
openai
pydantic
pandas
numpy
cryptography
